refactor: unify fieldInfo structure and streamline raw field handling (#36)
* refactor: unify fieldInfo structure and streamline raw field handling for improved performance

* chore: fix arg

* refactor: streamline boundary quote handling in SIMD scanning

* chore: fix lint error

* chore: fix format

* chore: fix format

+++ field_parser_test.go
@@ -1266,7 +1266,7 @@ func extractFieldContent(buf []byte, f fieldInfo) string {
 	if f.length == 0 {
 		return ""
 	}
-	if f.start+f.length > uint64(len(buf)) {
+	if f.start+f.length > uint32(len(buf)) {
 		return ""
 	}
 	return string(buf[f.start : f.start+f.length])
